[WIP] Hackathon merge (#3013)
Refactors workspace bootstrapping code:
Addition of projects volumes to ws-machines and 
applying of agents moved to infrastructure provisioner. 
It can be overwritten in any implementation of Che and 
will be responsible for providing all infrastructure needed for Che implementation functioning.
Allow to configure environment network driver.
Allow to add Agents from the code instead of special files. 
These agents can override agents configured in files if identifiers are equal.
Adds possibility to host agents binaries from CHE master.
Reworks agents in a way when they can be downloaded from CHE master if they are not found on FS.
Adds workspace ID to internal representation of environment to allow to use it for internal purposes of machine bootstrapping.
Renames some properties.
Refactors AgentConfigApplier to encapsulate additional code that was used with this class before.
Signed-off-by: Alexander Garagatyi <agaragatyi@codenvy.com>