Composer project generation (#3726)
* Composer project importer

Composer is a popular PHP dependency manager (similar to Bower for
Javascript). A natural way to create a new project in PHP is by
materializing it from a Composer package like this:

  composer create-project laravel/laravel /projects/laravel

The above command fetches the 'laravel/laravel' Composer package from a
central repository (https://packagist.org/) and installs it at the
/project/laravel folder. The advantage against cloning a git repository
is that `composer create-project` will install all required
dependencies.

This patch contributes a new Composer plugin that:
- Defines a Composer project importer that can import the project source
from a Composer package.
- Defines an Import Project Registrar that registers the Composer
importer into the import project wizard.

The Composer project importer uses the 'location' property of the
SourceStorage model for the package name. As result the "source" of
project templates can be defined like this:

  "source": {
    "type": "composer",
    "location": "zendframework/zend-expressive-skeleton",
    "parameters": {}
  }

Signed-off-by: Kaloyan Raev <kaloyan.r@zend.com>
Signed-off-by: Vitalii Parfonov <vparfonov@codenvy.com>