CLI - Fix error with version check on nightly (#3848)
With the new commit made where we check for newer versions on DockerHub, this check was being done for all versions including nightly. The compare_versions method would break if you had a nightly image. This fix adds a test to only do this check if you are not nightly.

Also, this cleans up some of the code around boot2docker checks. Unfortunately, since we are running our client inside of a Docker container, all we can do is print a warning message that you are on boot2docker and that you need to verify your host mounts.

Signed-off-by: Tyler Jewell tjewell@codenvy.com <tjewell@codenvy.com>